edit view-controller / user

+++ resources/views/admin/users/index.blade.php
@@ -18,8 +18,12 @@
             <table id="dataTable" class="table table-striped table-bordered" cellspacing="0" width="100%">
                 <thead>
                     <tr>
+                        <th>
+                            <input type="checkbox" id="CheckAll">
+                        </th>
                         <th>First name</th>
                         <th>Last name</th>
+                        <th>{{ trans('app.role') }}</th>
                         <th>Email</th>
                         <th>Phone number</th>
                         <th>Address</th>
@@ -29,8 +33,10 @@
 
                 <tfoot>
                     <tr>
+                        <th></th>
                         <th>First name</th>
                         <th>Last name</th>
+                        <th>{{ trans('app.role') }}</th>
                         <th>Email</th>
                         <th>Phone number</th>
                         <th>Address</th>
@@ -41,8 +47,20 @@
                 <tbody>
                     @foreach ($items as $item)
                         <tr>
+                            <td>
+                                @if($item->RolesStr !== App\Enums\UserRoles::SuperAdmin)
+                                        <div class="form-check">
+                                            <input type="checkbox" form="groupedAction" name="ids[]" value="{{ $item->id }}" class="form-check-input checkBoxClass">
+                                        </div>
+                                    @else
+                                            <i class="fa fa-ban text-danger" aria-hidden="true"></i>
+                                @endif
+                            </td>
                             <td><a href="{{ route(ADMIN . '.users.edit', $item->id) }}">{{ $item->first_name }}</a></td>
                             <td>{{ $item->last_name }}</td>
+                            <td>
+                                <label class="badge badge-success">{{ $item->RolesStr}}</label>
+                            </td>
                             <td>{{ $item->email }}</td>
                             <td>{{ $item->phone }}</td>
                             <td>{{ $item->address }}</td>
@@ -51,18 +69,22 @@
                                 <ul class="list-inline">
                                     <li class="list-inline-item">
                                         <a href="{{ route(ADMIN . '.users.edit', $item->id) }}" title="{{ trans('app.edit_title') }}" class="btn btn-primary btn-sm"><span class="ti-pencil"></span></a></li>
-                                    <li class="list-inline-item">
-                                        {!! Form::open([
-                                            'class'=>'delete',
-                                            'url'  => route(ADMIN . '.users.destroy', $item->id),
-                                            'method' => 'DELETE',
-                                            ])
-                                        !!}
+                                    @if($item->RolesStr !== App\Enums\UserRoles::SuperAdmin)
 
-                                            <button class="btn btn-danger btn-sm" title="{{ trans('app.delete_title') }}"><i class="ti-trash"></i></button>
+                                        <li class="list-inline-item">
+                                            {!! Form::open([
+                                                'class'=>'delete',
+                                                'url'  => route(ADMIN . '.users.destroy', $item->id),
+                                                'method' => 'DELETE',
+                                                ])
+                                            !!}
 
-                                        {!! Form::close() !!}
-                                    </li>
+                                                <button class="btn btn-danger btn-sm" title="{{ trans('app.delete_title') }}"><i class="ti-trash"></i></button>
+
+                                            {!! Form::close() !!}
+                                        </li>
+                                    @endif
+
                                 </ul>
                             </td>
                         </tr>
